CHG: added base test class (hidden in a wrapper class) w/ test utilities
CHG: moved lamps reader to data_readers module.
CHG: grouped all allegedly unused funcitions in data_readers in a class 'Deprecated'.
CHG: TestQEMDReader uses tests util to do first tests now.
